LPS-198979 Move getPartitionedName() methods to SchedulerEngine.java so that other modules have access to them
